// src/components/input/types.ts
export enum Inputs {
  INPUT_ADDON = 'InputAddon',
  INPUT_LEFT_ADDON = 'InputLeftAddon',
  INPUT_RIGHT_ADDON = 'InputRightAddon',
  INPUT_ELEMENT = 'InputElement',
  INPUT_LEFT_ELEMENT = 'InputLeftElement',
  INPUT_RIGHT_ELEMENT = 'InputRightElement',
  INPUT_CONTAINER = 'InputContainer',
  INPUT_LABEL = 'InputLabel',
  INPUT_GROUP = 'InputGroup',
  INPUT = 'Input',
}

export type Placement = 'left' | 'right'